+++ cutplace/__init__.py
@@ -6,11 +6,10 @@ Additionally to the command line tool the functionality of cutplace is also
 accessible through a Python API.
 """
 from cutplace.data import DataFormat, FORMAT_DELIMITED, FORMAT_EXCEL, FORMAT_FIXED, FORMAT_ODS
-from cutplace.errors import \
-    CheckError, CutplaceError, DataError, DataFormatError, FieldValueError, InterfaceError, Location, RangeValueError
+from cutplace.errors import Location
 from cutplace.interface import Cid
 from cutplace.ranges import Range
-from cutplace.validio import Reader
+from cutplace.validio import Reader, Writer
 from cutplace._version import get_versions
 
 #: Package version information.
@@ -20,20 +19,13 @@ del get_versions
 #: Public classes and functions.
 __all__ = [
     'Cid',
-    'CheckError',
-    'CutplaceError',
-    'DataError',
-    'DataFormat',
-    'DataFormatError',
     'FORMAT_DELIMITED',
     'FORMAT_EXCEL',
     'FORMAT_FIXED',
     'FORMAT_ODS',
-    'FieldValueError',
-    'InterfaceError',
     'Location',
     'Range',
-    'RangeValueError',
     'Reader',
+    'Writer',
     '__version__'
 ]