+++ cutplace/interface.py
@@ -101,7 +101,9 @@ class Cid(object):
     def data_format(self):
         """
         The data format used by the this CID; refer to the
-        py:mod:`cutplace.data` module for possible formats.
+        :py:mod:`cutplace.data` module for possible formats.
+
+        :rtype: cutplace.data.DataFormat
         """
         return self._data_format
 
@@ -265,6 +267,7 @@ class Cid(object):
             self._location.advance_line()
         if self.data_format is None:
             raise errors.InterfaceError('data format must be specified', self._location)
+        self.data_format.validate()
         if len(self.field_names) == 0:
             raise errors.InterfaceError('fields must be specified', self._location)
 
@@ -532,6 +535,24 @@ def create_cid_from_string(cid_text):
     """
     with io.StringIO(cid_text) as cid_string_io:
         result = Cid(cid_string_io)
+    return result
+
+
+def field_lengths(fixed_cid):
+    """
+    List of :py:class:`int`s for all field lengths in ``fixed_cid`` which
+    must be of data format py:attr:`~cutplace.data.FORMAT_FIXED`.
+    """
+    assert fixed_cid is not None
+    assert fixed_cid.data_format.format == data.FORMAT_FIXED, 'format=' + fixed_cid.data_format.format
+    result = []
+    for field_format in fixed_cid.field_formats:
+        field_length_range = field_format.length.items[0]
+        lower, upper = field_length_range
+        assert lower is not None
+        assert lower == upper
+        field_length = lower
+        result.append(field_length)
     return result